requirements.txt

flask
gunicorn
transformers
torch
sentencepiece
newspaper3k
lxml_html_clean
